# services/wallet/internal/model/generated.go
//go:build go1.22

// Package model provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package model

import (
	"errors"
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for TransactionResponseStatus.
const (
	Completed TransactionResponseStatus = "completed"
	Failed    TransactionResponseStatus = "failed"
	Pending   TransactionResponseStatus = "pending"
)

// Valid indicates whether the value is a known member of the TransactionResponseStatus enum.
func (e TransactionResponseStatus) Valid() bool {
	switch e {
	case Completed:
		return true
	case Failed:
		return true
	case Pending:
		return true
	default:
		return false
	}
}

// Defines values for TransactionResponseType.
const (
	Bet        TransactionResponseType = "bet"
	Deposit    TransactionResponseType = "deposit"
	Payout     TransactionResponseType = "payout"
	Transfer   TransactionResponseType = "transfer"
	Withdrawal TransactionResponseType = "withdrawal"
)

// Valid indicates whether the value is a known member of the TransactionResponseType enum.
func (e TransactionResponseType) Valid() bool {
	switch e {
	case Bet:
		return true
	case Deposit:
		return true
	case Payout:
		return true
	case Transfer:
		return true
	case Withdrawal:
		return true
	default:
		return false
	}
}

// AccountResponse defines model for AccountResponse.
type AccountResponse struct {
	AccountType *string `json:"accountType,omitempty"`

	// Balance Balance as decimal string (supports fiat 2dp and crypto up to 18dp)
	Balance  *string `json:"balance,omitempty"`
	Currency *string `json:"currency,omitempty"`
	UserId   *int64  `json:"userId,omitempty"`
}

// DepositRequest defines model for DepositRequest.
type DepositRequest struct {
	// Amount Amount as decimal string (supports fiat 2dp and crypto up to 18dp), must be positive
	Amount   string `json:"amount"`
	Currency string `json:"currency"`

	// Source External payment reference (PSP transaction id)
	Source *string `json:"source,omitempty"`
	UserId int64   `json:"userId"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code      *int       `json:"code,omitempty"`
	Error     *string    `json:"error,omitempty"`
	Timestamp *time.Time `json:"timestamp,omitempty"`
}

// TransactionResponse defines model for TransactionResponse.
type TransactionResponse struct {
	// Amount Amount as decimal string (supports fiat 2dp and crypto up to 18dp)
	Amount    *string                    `json:"amount,omitempty"`
	CreatedAt *time.Time                 `json:"createdAt,omitempty"`
	Currency  *string                    `json:"currency,omitempty"`
	Id        *openapi_types.UUID        `json:"id,omitempty"`
	Status    *TransactionResponseStatus `json:"status,omitempty"`
	Type      *TransactionResponseType   `json:"type,omitempty"`
}

// TransactionResponseStatus defines model for TransactionResponse.Status.
type TransactionResponseStatus string

// TransactionResponseType defines model for TransactionResponse.Type.
type TransactionResponseType string

// TransferRequest defines model for TransferRequest.
type TransferRequest struct {
	// Amount Amount as decimal string (supports fiat 2dp and crypto up to 18dp)
	Amount     string `json:"amount"`
	Currency   string `json:"currency"`
	FromUserId int64  `json:"fromUserId"`

	// Reason e.g. 'bet_placed', 'bet_payout'
	Reason string `json:"reason"`

	// ToAccountType e.g. 'system_house' for bet placement
	ToAccountType string `json:"toAccountType"`
}

// WithdrawRequest defines model for WithdrawRequest.
type WithdrawRequest struct {
	// Amount Amount as decimal string (supports fiat 2dp and crypto up to 18dp)
	Amount   string `json:"amount"`
	Currency string `json:"currency"`
	UserId   int64  `json:"userId"`
}

// GetAccountParams defines parameters for GetAccount.
type GetAccountParams struct {
	Currency *string `form:"currency,omitempty" json:"currency,omitempty"`
}

// DepositParams defines parameters for Deposit.
type DepositParams struct {
	// IdempotencyKey Client-generated UUID to prevent duplicate processing
	IdempotencyKey string `json:"Idempotency-Key"`
}

// ListTransactionsParams defines parameters for ListTransactions.
type ListTransactionsParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// TransferParams defines parameters for Transfer.
type TransferParams struct {
	IdempotencyKey string `json:"Idempotency-Key"`
}

// WithdrawParams defines parameters for Withdraw.
type WithdrawParams struct {
	IdempotencyKey string `json:"Idempotency-Key"`
}

// DepositJSONRequestBody defines body for Deposit for application/json ContentType.
type DepositJSONRequestBody = DepositRequest

// TransferJSONRequestBody defines body for Transfer for application/json ContentType.
type TransferJSONRequestBody = TransferRequest

// WithdrawJSONRequestBody defines body for Withdraw for application/json ContentType.
type WithdrawJSONRequestBody = WithdrawRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get user's account balance
	// (GET /wallet/accounts/{userId})
	GetAccount(w http.ResponseWriter, r *http.Request, userId int64, params GetAccountParams)
	// Deposit funds into user account
	// (POST /wallet/deposit)
	Deposit(w http.ResponseWriter, r *http.Request, params DepositParams)
	// Get transaction history for user
	// (GET /wallet/transactions/{userId})
	ListTransactions(w http.ResponseWriter, r *http.Request, userId int64, params ListTransactionsParams)
	// Internal transfer between accounts (e.g. bet placement, payout)
	// (POST /wallet/transfer)
	Transfer(w http.ResponseWriter, r *http.Request, params TransferParams)
	// Withdraw funds from user account
	// (POST /wallet/withdraw)
	Withdraw(w http.ResponseWriter, r *http.Request, params WithdrawParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAccount operation middleware
func (siw *ServerInterfaceWrapper) GetAccount(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "userId" -------------
	var userId int64

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetAccountParams

	// ------------- Optional query parameter "currency" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "currency", r.URL.Query(), &params.Currency, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "currency"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "currency", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAccount(w, r, userId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Deposit operation middleware
func (siw *ServerInterfaceWrapper) Deposit(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params DepositParams

	headers := r.Header

	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Idempotency-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Idempotency-Key", Err: err})
			return
		}

		params.IdempotencyKey = IdempotencyKey

	} else {
		err := fmt.Errorf("Header parameter Idempotency-Key is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Idempotency-Key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Deposit(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListTransactions operation middleware
func (siw *ServerInterfaceWrapper) ListTransactions(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "userId" -------------
	var userId int64

	err = runtime.BindStyledParameterWithOptions("simple", "userId", r.PathValue("userId"), &userId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "int64"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "userId", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params ListTransactionsParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "limit"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "offset"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListTransactions(w, r, userId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Transfer operation middleware
func (siw *ServerInterfaceWrapper) Transfer(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params TransferParams

	headers := r.Header

	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Idempotency-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Idempotency-Key", Err: err})
			return
		}

		params.IdempotencyKey = IdempotencyKey

	} else {
		err := fmt.Errorf("Header parameter Idempotency-Key is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Idempotency-Key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Transfer(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Withdraw operation middleware
func (siw *ServerInterfaceWrapper) Withdraw(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params WithdrawParams

	headers := r.Header

	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey string
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandlerFunc(w, r, &TooManyValuesForParamError{ParamName: "Idempotency-Key", Count: n})
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "Idempotency-Key", Err: err})
			return
		}

		params.IdempotencyKey = IdempotencyKey

	} else {
		err := fmt.Errorf("Header parameter Idempotency-Key is required, but not found")
		siw.ErrorHandlerFunc(w, r, &RequiredHeaderError{ParamName: "Idempotency-Key", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Withdraw(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/wallet/accounts/{userId}", wrapper.GetAccount)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/wallet/deposit", wrapper.Deposit)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/wallet/transactions/{userId}", wrapper.ListTransactions)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/wallet/transfer", wrapper.Transfer)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/wallet/withdraw", wrapper.Withdraw)

	return m
}